requirements/base.txt
pytz==2018.9  # https://github.com/stub42/pytz
python-slugify==2.0.1  # https://github.com/un33k/python-slugify
Pillow==5.4.1  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==18.3.0  # https://github.com/hynek/argon2_cffi

# Django
# ------------------------------------------------------------------------------
django==2.1.8
django-environ>=0.4.5  # https://github.com/joke2k/django-environ
django-model-utils>=3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth>=0.38.0  # https://github.com/pennersr/django-allauth
django-foundation-formtags>=0.1.0
django-compressor>=2.2  # https://github.com/django-compressor/django-compressor
django-libsass>=0.7
django-redis>=4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework>=3.9.0  # https://github.com/encode/django-rest-framework
coreapi>=2.3.3  # https://github.com/core-api/python-client


wagtail>=2.4
wagtailfontawesome>=1.1.3
celery>=4.2.1
django-wagtail-feeds>=0.1.0
libsass>=0.14.5
psycopg2-binary>=2.7.6.1 
wand>=0.5.0
wagtail-markdown>=0.5
bleach>=3.0.2
wagalytics>=0.8.1 
sentry-sdk